# api/openapi.yaml
openapi: 3.1.0
info:
  title: Catalift API
  version: 1.7.0
  description: |
    The REST API the Catalift web app uses: sellers upload products and start generation, reviewers edit,
    regenerate, approve and export listings, and both see AI cost and the budget state (ADR-0007).
    Serves: US-00-001 to US-00-012. Style: docs/api/README.md. The server default for X-API-Version is the
    newest date version, currently 2026-10-04.
  contact:
    name: Catalift team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-api-style: >-
    REST over HTTPS with JSON bodies, plus multipart for CSV and image uploads and text/csv for export
    downloads; one Go service exposing five areas (auth, catalogue, listings, export, cost) under one /v1.
  x-versioning: >-
    Major in the path (/v1, ADR-0007); dated behaviour changes inside a major in X-API-Version, server
    default the newest (2026-10-04).
  x-authentication: >-
    A server-side session in an HTTP-only secure cookie (catalift_session) set by POST /v1/sessions, with
    an X-CSRF-Token header on every state-changing request (ADR-0006). Roles seller and reviewer are
    checked on every route; a seller calling a reviewer action gets 403 forbidden_role.
  x-conventions:
    - rule: 'The base path is /v1, not /api/v1.'
      why: 'ADR-0007 and HLD section 5 fixed /v1; the reverse proxy forwards /v1 and serves the web app at /.'
...
    - rule: 'Ids are decimal strings of the database''s bigint ids.'
...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      summary: Configured channels, their rules and the last re-check
      description: >-
        Channels come from configuration files (REQ-017). A channel whose file failed to load is listed disabled with
        the error (D8); the last config re-check per channel carries the counts the grid shows (D7).
      tags: [channels]
      x-story-ids: [US-00-005, US-00-006, US-00-004]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/UploadFilter'
      responses:
        '200':
          description: One page of channels
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChannelPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /channels/{channel}:
    parameters:
      - name: channel
        in: path
        required: true
        schema: { type: string, pattern: '^[a-z][a-z0-9_]{1,39}$' }
    patch:
      operationId: updateChannelRules
      summary: Change a channel's rules (reviewer)
      description: >-
        Saves a reviewer's change to the title length limit, required attributes and banned words, records who made
        it, and re-checks the channel's listings at once; an approved listing that now fails loses its approval (D7).
        The body carries the config_hash from GET /v1/channels; a change made by someone else meanwhile is 409
        version_conflict (v1.3).
      tags: [channels]
      x-story-ids: [US-00-005, US-00-004]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChannelRulesUpdate' }
      responses:
        '200':
          description: The re-check the change caused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChannelRulesResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /exports:
    get:
      operationId: listExports
      summary: List exports, newest first
      description: >-
        A reviewer sees every export, or only sent ones with filter[sent]=true. A seller always sees only the exports
        a reviewer sent them (v1.6, ADR-0012).
      tags: [exports]
      x-story-ids: [US-00-010]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/UploadFilter'
        - name: filter[sent]
          in: query
          schema: { type: string, enum: ['true'] }
      responses:
        '200':
          description: One page of exports
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createExport
      summary: Export approved listings (reviewer)
      description: >-
        Writes one CSV per channel holding only listings approved at their current version, with headers from the
        channel's configuration and cells starting with =, +, - or @ escaped. A channel with no approved listings
        gets no file and is listed in skipped_channels (AC-US-00-010-5). With a body naming upload_id, only that
        upload's listings are exported; without a body, every upload's (v1.7).
      tags: [exports]
      x-story-ids: [US-00-010]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                upload_id: { type: string, pattern: '^[0-9]+$' }
            example: { upload_id: '12' }
      responses:
        '201':
          description: The export and its files
          headers:
            Location: { schema: { type: string, format: uri } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Export' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /exports/{export_id}:
    parameters:
      - $ref: '#/components/parameters/ExportId'
    get:
      operationId: getExport
      summary: Get an export and its files
      description: A seller gets 404 for an export not yet sent to them (v1.6).
      tags: [exports]
      x-story-ids: [US-00-010]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The export
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Export' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /exports/{export_id}/send:
    parameters:
      - $ref: '#/components/parameters/ExportId'
    post:
      operationId: sendExport
      summary: Send an export to the seller (reviewer)
      description: >-
        Marks the export as sent, with who sent it and when, so it appears in the seller's received exports and the
        seller can download its files. Sending again changes nothing and returns the first send (v1.6, ADR-0012).
      tags: [exports]
      x-story-ids: [US-00-010]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: The sent export
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Export' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /exports/{export_id}/files/{channel}:
    parameters:
      - $ref: '#/components/parameters/ExportId'
      - name: channel
        in: path
        required: true
        schema: { type: string, pattern: '^[a-z][a-z0-9_]{1,39}$' }
    get:
      operationId: getExportFile
      summary: Download one channel's CSV
      description: Reviewers download any export; a seller only one sent to them, otherwise 404 (v1.6).
      tags: [exports]
      x-story-ids: [US-00-010]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The CSV file
          headers:
            Content-Disposition: { schema: { type: string }, description: 'attachment; filename="<channel>-<export_id>.csv"' }
          content:
            text/csv:
              schema: { type: string }
              example: "sku,title,bullet_1,bullet_2,bullet_3,bullet_4,bullet_5,description,colour,pattern,sleeve,neckline,fit\nKU-104,Women's Black Cotton Kurta,...\n"
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /budget:
    get:
      operationId: getBudget
      summary: AI spend, the limit and whether calls are blocked
      description: >-
        spent_micro_usd uses the one spend expression of the data model (succeeded at cost, failed at cost or 0,
        reserved and possibly_charged at their reservation). blocked_at drives the banner (D13).
      tags: [budget]
      x-story-ids: [US-00-011, US-00-012]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The budget state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: catalift_session
      description: Server-side session set by POST /v1/sessions (ADR-0006).
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      description: Date version, e.g. 2026-10-04. Missing means the server default (the newest).
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client UUID. Same key and body replays the first result for 24 hours; same key and a different body is 409 idempotency_conflict.
      schema: { type: string, format: uuid }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: The csrf_token from the session; required on every state-changing request (ADR-0006).
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from page.next_cursor.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    UploadFilter:
      name: filter[upload_id]
      in: query
      description: Only this upload's records; the screens pass the latest upload (v1.7).
      schema: { type: string, pattern: '^[0-9]+$' }
...
    ExportId:
      name: export_id
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9]+$' }
...
  schemas:
    Error:
      type: object
      required: [error]
      x-error-codes:
        - { code: validation_failed, status: 400, meaning: "Malformed request, an unknown filter, or a field failed validation; details names the field." }
        - { code: unauthorized, status: 401, meaning: "No session, an expired one, or wrong email or password at sign-in." }
        - { code: forbidden_role, status: 403, meaning: "Signed in, but this action needs the reviewer role (AC-US-00-009-5)." }
        - { code: csrf_failed, status: 403, meaning: "X-CSRF-Token missing or not this session's token." }
        - { code: not_found, status: 404, meaning: "No such record." }
        - { code: idempotency_conflict, status: 409, meaning: "Idempotency-Key reused with a different body." }
        - { code: version_conflict, status: 409, meaning: "The listing or attributes changed since the version sent; reload and retry." }
        - { code: payload_too_large, status: 413, meaning: "CSV over 5 MB or an image over 10 MB." }
        - { code: unsupported_media_type, status: 415, meaning: "Not CSV, or an image that is not JPEG, PNG or WebP." }
        - { code: unprocessable, status: 422, meaning: "Valid shape but the content breaks a rule; details names it." }
        - { code: voice_note_required, status: 422, meaning: "A brand in scope has no voice note and neutral_voice_confirmed is false (AC-US-00-003-5)." }
        - { code: budget_blocked, status: 422, meaning: "AI calls are blocked until the owner clears the budget block (D13)." }
        - { code: listing_not_generated, status: 422, meaning: "The listing has no generated text yet, so a field cannot be regenerated." }
        - { code: no_approved_listings, status: 422, meaning: "No channel has an approved listing, so no file would be written (AC-US-00-010-5)." }
        - { code: rate_limited, status: 429, meaning: "Too many requests; wait Retry-After seconds." }
        - { code: internal, status: 500, meaning: "Unexpected failure; quote request_id to support." }
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  reason: { type: string }
            request_id: { type: string }
      example:
        error:
          code: voice_note_required
          message: Brand Indigo Loom has no voice note; add one or confirm a neutral voice.
          details: [{ field: neutral_voice_confirmed, reason: must be true when a brand has no voice note }]
          request_id: req_01J9Z6
    Page:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
      example: { next_cursor: eyJza3UiOiJLVS0xMDQifQ, has_more: true }
...
    Channel:
      type: object
      required: [id, name, enabled, load_error, title_max_length, required_attributes, banned_words, export_headers, last_recheck, listings_total, listings_approved]
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        load_error: { type: [string, 'null'], description: Why the file failed to load; the channel is disabled (D8). }
        title_max_length: { type: [integer, 'null'] }
        required_attributes: { type: array, items: { type: string } }
        banned_words: { type: array, items: { type: string }, description: Words and phrases the rules engine refuses (v1.1). }
        config_hash:
          type: [string, 'null']
          description: Identifies the rules in force; send it back with PATCH /v1/channels/{channel} (v1.3). Null for a disabled channel.
        last_edit:
          type: [object, 'null']
          description: The last change made in the app, if any; the channel file's rules apply otherwise (v1.3).
          required: [by, at]
          properties:
            by: { type: string, description: Email of the reviewer. }
            at: { type: string, format: date-time }
        listings_total: { type: integer, description: Listings on this channel (v1.1). }
        listings_approved: { type: integer, description: Listings approved at their current version; the export page shows readiness from these (v1.1). }
        export_headers: { type: array, items: { type: string } }
        last_recheck:
          type: [object, 'null']
          required: [config_hash, listings_rechecked, approvals_cleared, created_at]
          properties:
            config_hash: { type: string }
            listings_rechecked: { type: integer }
            approvals_cleared: { type: integer }
            created_at: { type: string, format: date-time }
      example:
        id: amazon_style
        name: Amazon-style marketplace
        enabled: true
        load_error: null
        title_max_length: 200
        required_attributes: [colour, fit]
        banned_words: [best seller, guaranteed]
        listings_total: 300
        listings_approved: 296
        export_headers: [sku, item_name, bullet_point_1, bullet_point_2, bullet_point_3, bullet_point_4, bullet_point_5, product_description, color, fit_type]
        last_recheck: { config_hash: 9b1f2c, listings_rechecked: 300, approvals_cleared: 4, created_at: 2026-10-04T07:00:00Z }
    ChannelRulesUpdate:
      type: object
      required: [config_hash, title_max_length, required_attributes, banned_words]
      properties:
        config_hash: { type: string, description: The channel's config_hash the reviewer edited. }
        title_max_length: { type: integer, minimum: 10, maximum: 500 }
        required_attributes:
          type: array
          items: { type: string, enum: [colour, pattern, sleeve, neckline, fit] }
        banned_words: { type: array, maxItems: 200, items: { type: string, maxLength: 100 } }
      example: { config_hash: 9b1f2c3d4e5f6a7b, title_max_length: 150, required_attributes: [colour, fit], banned_words: [best seller, guaranteed] }
    ChannelRulesResult:
      type: object
      required: [channel, listings_rechecked, approvals_cleared]
      properties:
        channel: { type: string }
        listings_rechecked: { type: integer, minimum: 0 }
        approvals_cleared: { type: integer, minimum: 0 }
      example: { channel: own_website, listings_rechecked: 300, approvals_cleared: 4 }
    ChannelPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Channel' } }
        page: { $ref: '#/components/schemas/Page' }
      example:
        data:
          - { id: own_website, name: Own website, enabled: true, load_error: null, title_max_length: 120, required_attributes: [colour], export_headers: [sku, title, bullet_1, bullet_2, bullet_3, bullet_4, bullet_5, description], last_recheck: null }
        page: { next_cursor: null, has_more: false }
    Export:
      type: object
      required: [id, created_at, files, skipped_channels, sent_at, sent_by]
      properties:
        id: { type: string }
        created_at: { type: string, format: date-time }
        sent_at: { type: [string, 'null'], format: date-time, description: When a reviewer sent it to the seller. }
        sent_by: { type: [string, 'null'], description: Email of the reviewer who sent it. }
        files:
          type: array
          items:
            type: object
            required: [channel, row_count, download_url]
            properties:
              channel: { type: string }
              row_count: { type: integer, minimum: 1 }
              download_url: { type: string }
        skipped_channels:
          type: array
          items:
            type: object
            required: [channel, reason]
            properties:
              channel: { type: string }
              reason: { type: string }
      example:
        id: '9'
        created_at: 2026-10-04T11:00:00Z
        files: [{ channel: amazon_style, row_count: 298, download_url: /v1/exports/9/files/amazon_style }]
        skipped_channels: [{ channel: own_website, reason: no approved listings }]
        sent_at: null
        sent_by: null
    ExportPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Export' } }
        page: { $ref: '#/components/schemas/Page' }
      example:
        data:
          - id: '9'
            created_at: 2026-10-04T11:00:00Z
            files: [{ channel: amazon_style, row_count: 298, download_url: /v1/exports/9/files/amazon_style }]
            skipped_channels: []
            sent_at: 2026-10-04T11:05:00Z
            sent_by: reviewer@example.com
        page: { next_cursor: null, has_more: false }
    Budget:
      type: object
      required: [limit_micro_usd, spent_micro_usd, blocked_at]
      properties:
        limit_micro_usd: { type: integer, minimum: 1 }
        spent_micro_usd: { type: integer, minimum: 0 }
        blocked_at: { type: [string, 'null'], format: date-time, description: Set on the first refused call; the banner shows while it is set. }
      example: { limit_micro_usd: 8000000, spent_micro_usd: 4210000, blocked_at: null }
...
  responses:
    BadRequest:
      description: Malformed request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: No session, or wrong email or password
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Wrong role or missing CSRF token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such record
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency key reused with a different body, or a stale version
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    UnprocessableEntity:
      description: Valid shape, but the content breaks a rule
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    InternalError:
      description: Unexpected failure; request_id is in the body
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }